refactor: 게임참여자 추상화 객체 Participant 구현
딜러와 플레이어는 Participant 상속받음
