+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,19 +1,16 @@
 package blackjack.domain;
 
-import java.util.List;
 import java.util.function.BooleanSupplier;
 
-public class Dealer {
+public class Dealer extends Participant {
+
     private static final int DEALER_DRAW_THRESHOLD = 17;
 
-    private final Name name;
-    private final Hands hands;
-
     public Dealer(String name) {
-        this.name = new Name(name);
-        this.hands = new Hands();
+        super(name);
     }
 
+    @Override
     public void decideDraw(BooleanSupplier supplier, Deck deck) {
         if (supplier.getAsBoolean()) {
             Card card = deck.draw();
@@ -23,13 +20,5 @@ public class Dealer {
 
     public Boolean shouldDraw() {
         return hands.getHandsScore() < DEALER_DRAW_THRESHOLD;
-    }
-
-    public void addCard(Card card) {
-        hands.addCard(card);
-    }
-
-    public List<Card> getHandsCards() {
-        return hands.getHands();
     }
 }
